fix: address PR #391 review comments
- Update CBOR key ordering comment in messages.rs to note KEY_PAYLOAD
  is optional for Nop/Reboot commands
- Move non-T-Pxxx tests (key_hint_from_psk, CBOR key ordering) from
  validation.rs to a new tests/unit.rs to preserve traceability
